refactor(web): improve file structure and layout
