+++ web/src/layouts/settings/index.tsx
@@ -1,0 +1,9 @@
+const Settings: React.FC = () => {
+  return (
+    <>
+      <>Hello there</>
+    </>
+  );
+};
+
+export default Settings;
